pull all the file helper classes out into a different file
