+++ polling_stations/apps/data_collection/management/commands/__init__.py
@@ -2,23 +2,17 @@
 Defines the base importer classes to implement
 """
 import abc
-import csv
 import json
 import glob
 import os
 import re
-import shapefile
 import tempfile
 import unicodedata
 import urllib.request
-import zipfile
-
-from collections import namedtuple
 
 from django.core.management.base import BaseCommand
 from django.conf import settings
 from django.contrib.gis import geos
-from django.contrib.gis.gdal import DataSource, GDALException
 from django.contrib.gis.geos import Point, GEOSGeometry
 from django.db import connection
 from django.db import transaction
@@ -32,6 +26,7 @@ from data_collection.data_quality_report import (
     DistrictReport,
     ResidentialAddressReport
 )
+from data_collection.filehelpers import FileHelperFactory
 from pollingstations.models import (
     PollingStation,
     PollingDistrict,
@@ -39,114 +34,6 @@ from pollingstations.models import (
 )
 from data_collection.models import DataQuality
 from addressbase.helpers import create_address_records_for_council
-
-
-class CsvHelper:
-
-    def __init__(self, filepath, encoding='utf-8', delimiter=','):
-        self.filepath = filepath
-        self.encoding = encoding
-        self.delimiter = delimiter
-
-    def parseCsv(self):
-        file = open(self.filepath, 'rt', encoding=self.encoding)
-        reader = csv.reader(file, delimiter=self.delimiter)
-        header = next(reader)
-
-        # mimic the data structure generated by ffs so existing import
-        # scripts don't break
-        replace = {
-            ' ': '_',
-            '-': '_',
-            '.': '_',
-            '(': '',
-            ')': '',
-        }
-        clean = []
-        for s in header:
-            s = s.strip().lower()
-            for k, v in replace.items():
-                s = s.replace(k, v)
-            while '__' in s:
-                s = s.replace('__', '_')
-            clean.append(s)
-        RowKlass = namedtuple('RowKlass', clean)
-
-        data = []
-        for row in map(RowKlass._make, reader):
-            data.append(row)
-
-        file.close()
-        return data
-
-    # convenience alias for parseCsv()
-    def get_features(self):
-        return self.parseCsv()
-
-
-class ShpHelper:
-
-    def __init__(self, filepath):
-        self.filepath = filepath
-
-    def get_features(self):
-        sf = shapefile.Reader(self.filepath)
-        return sf.shapeRecords()
-
-
-class JsonHelper:
-
-    def __init__(self, filepath):
-        self.filepath = filepath
-
-    def get_features(self):
-        geometies = json.load(open(self.filepath))
-        return geometies['features']
-
-
-class KmlHelper:
-
-    def __init__(self, filepath):
-        self.filepath = filepath
-
-    def parse_features(self, kml):
-        try:
-            ds = DataSource(kml)
-        except GDALException:
-            # This is very strange – sometimes the above will fail the first
-            # time, but not the second. Seen on OS X with GDAL 2.2.0
-            ds = DataSource(kml)
-        return ds[0]
-
-    def get_features(self):
-        if not self.filepath.endswith('.kmz'):
-            return self.parse_features(self.filepath)
-
-        # It's a .kmz file
-        # Because the C lib that the Django DataSource is wrapping
-        # expects a file on disk, let's extract the KML to a tmpfile
-        kmz = zipfile.ZipFile(self.filepath, 'r')
-        kmlfile = kmz.open('doc.kml', 'r')
-
-        with tempfile.NamedTemporaryFile() as tmp:
-            tmp.write(kmlfile.read())
-            data = self.parse_features(tmp.name)
-            tmp.close()
-            return data
-
-
-class FileHelperFactory():
-
-    @staticmethod
-    def create(filetype, filepath, options):
-        if (filetype == 'shp'):
-            return ShpHelper(filepath)
-        elif (filetype == 'kml'):
-            return KmlHelper(filepath)
-        elif (filetype == 'json'):
-            return JsonHelper(filepath)
-        elif (filetype == 'csv'):
-            return CsvHelper(filepath, options['encoding'], options['delimiter'])
 
 
 class Database:
